zmiana sposobu wybierania trasy. Brak dodania obslugi feromonów.

+++ semestr_6/rozproszone/projekt_01/src/droga.h
@@ -7,7 +7,7 @@ class Droga
 		void DodajFeromony();
 //	private:
 		int itsDl;
-		int itsFeromony;
+		double itsFeromony;
 		int itsMiasta[2];
 };
 #endif
